requirements.txt
Django==1.9.2
django-polymorphic==0.9.1
djangorestframework==3.5.1
django-rest-swagger==2.1.0
drf-extensions==0.2.8
drf-nested-routers==0.11.1
PyYAML==3.11
wheel==0.29.0
requests==2.9.1
multimethod==0.5
behave-django==0.3.0

# Development
django-extensions==1.6.1
Werkzeug==0.11.3
sphinx>=1.4.0
sphinx-rtd-theme


-e git+https://github.com/glibersat/brewpi.service.devices.core#egg=brewpi.service.devices.core
